# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/cj/reocrd/view/activity/AddressActivity.java
#	app/src/main/java/com/cj/reocrd/view/activity/OrderDetailActivity.java
#	app/src/main/res/layout/activity_order_detail.xml

+++ app/src/main/java/com/cj/reocrd/view/activity/WalletGetActivity.java
@@ -57,7 +57,7 @@ public class WalletGetActivity extends BaseActivity<MyPrresenter> implements MyC
     TextView tvAlipayName;
     @BindView(R.id.tvBtnBind)
     TextView tvBtnBind;
-    int useableblance;
+    Double useableblance;
     MyBankAdapter adapter;
     List<BankBean.Bank> banks;
     BankBean.Bank bank;
@@ -75,7 +75,7 @@ public class WalletGetActivity extends BaseActivity<MyPrresenter> implements MyC
         super.initData();
         aliPayName = (String) SPUtils.get(this,SPUtils.SpKey.ALIPAY_NAME,"");
         Bundle bundle = getIntent().getExtras();
-        useableblance = bundle.getInt("useableblance");
+        useableblance = bundle.getDouble("useableblance");
         type = 1;
         mPresenter.myCard(UrlConstants.UrLType.MY_CARD, uid);
     }
